[REFACTOR] Rework screen resolution into service
