+++ pandora-client-web/src/services/clientServices.ts
@@ -5,12 +5,14 @@ import { DirectMessageManagerServiceProvider, type DirectMessageManager } from '
 import { AudioServiceProvider, type AudioService } from './audio.ts';
 import { BrowserPermissionManagerServiceProvider, type BrowserPermissionManager } from './browserPermissionManager.ts';
 import { NotificationHandlerServiceProvider, type NotificationHandler } from './notificationHandler.tsx';
+import { ScreenResolutionServiceProvider, type ScreenResolutionService } from './screenResolution/screenResolution.ts';
 import { ShardConnectionManagerServiceProvider, type IShardConnectionManager } from './shardConnectionManager.ts';
 import { UserActivationServiceProvider, type UserActivationService } from './userActivation.ts';
 
 /** Services available on Padora's client, when running in normal user mode. */
 export type ClientServices = Satisfies<
 	{
+		screenResolution: ScreenResolutionService;
 		browserPermissionManager: BrowserPermissionManager;
 		userActivation: UserActivationService;
 		audio: AudioService;
@@ -28,6 +30,7 @@ export type ClientServices = Satisfies<
  */
 export function GenerateClientUsermodeServices(): ServiceManager<ClientServices> {
 	return new ServiceManager<ClientServices>()
+		.registerService(ScreenResolutionServiceProvider)
 		.registerService(BrowserPermissionManagerServiceProvider)
 		.registerService(UserActivationServiceProvider)
 		.registerService(AudioServiceProvider)
